clarify the decision for this project's approach
Previously it was not clear what benefits this project's approach
offered over simply tracking all the changes in Git, and building
tarballs directly from a Git branch.  This should make it a bit
clearer, although it would be nice if some further details about this
use case were added at some point in the future.

Closes #1 and #2.