Handle non-UTF-8 bytes in patches
The approach I took was to "surrogateescape" non-UTF-8 bytes. When the
patch files are opened to search for the skip tag, non-UTF-8 bytes
aren't interesting since the skip tag is valid UTF-8 in almost all
cases. The other situation is manually copying the patch files on
`--update`, here it's required to not mess with non-UTF-8 bytes inside
the patches.